requirements.txt
blinker==1.9.0
cachelib==0.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cryptography==46.0.5
DBUtils==3.0.3
dicttoxml==1.7.16
Flask==3.0.2
Flask-Caching==2.1.0
gunicorn==21.2.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pluggy==1.6.0
pycparser==3.0
Pygments==2.19.2
PyJWT==2.8.0
PyMySQL==1.1.0
pytest==9.0.2
pytest-flask==1.3.0
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.6.3
Werkzeug==3.1.6